Made new offence ui invisible
And commented out the reading in part.

+++ AI_WallDef_Editor/AIEdit.cs
@@ -47,7 +47,7 @@ namespace AI_WallDef_Editor
         public List<AttackingUnit> Attackers = new List<AttackingUnit>();
         public int FirstAttackForceSize;
         public int AttackForceIncreaseStatic; 
-        public int AttackForceIncreaseMultiplier;
+        public double AttackForceIncreaseMultiplier;
     }
